backend/requirements.txt
# Core Framework
fastapi==0.121.2
uvicorn==0.38.0

# Database
motor==3.7.1
pymongo==4.15.4

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0

# Data Validation
pydantic==2.12.4
pydantic[email]==2.12.4
email-validator==2.3.0

# HTTP & API
httpx==0.28.1
python-multipart==0.0.20

# AI/LLM
groq>=0.4.0

# Environment
python-dotenv==1.2.1
